requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
openai>=2.0.0
Werkzeug==3.0.1

